osd: change warn_interval_multiplier to uint32_t

+++ src/osd/OpRequest.h
@@ -135,7 +135,7 @@ struct OpRequest : public TrackedOp {
   void set_pg_op() { rmw_flags |= CEPH_OSD_RMW_FLAG_PGOP; }
 
   utime_t received_time;
-  uint8_t warn_interval_multiplier;
+  uint32_t warn_interval_multiplier;
   utime_t get_arrived() const {
     return received_time;
   }
